+++ core/forms.py
@@ -1,3 +1,5 @@
+import json
+
 from directory_components import forms
 from directory_constants import choices
 from directory_forms_api_client.forms import GovNotifyEmailActionMixin
@@ -170,16 +172,19 @@ class ProductSearchForm(forms.Form):
         container_css_classes='js-enabled-only',
         widget=TextInput(attrs={'form': 'search-form'}),
     )
-    commodities = forms.CharField(
+    commodity = forms.CharField(
         label='Commodity codes',
-        help_text='Find the commodity codes via the commodity code browser. Comma separated.',
+        help_text='Find the commodity codes via the commodity code browser.',
         widget=HiddenInput,
     )
 
     def clean(self):
         super().clean()
-        if not self.cleaned_data.get('commodities'):
+        if not self.cleaned_data.get('commodity'):
             self.add_error('term', self.MESSAGE_MISSING_PRODUCT)
+
+    def clean_commodity(self):
+        return json.loads(self.cleaned_data['commodity'])
 
 
 class OtherMetricNameForm(forms.Form):
@@ -468,3 +473,7 @@ class EquivalendUKGoodsForm(fields.BindNestedFormMixin, forms.Form):
         coerce=lambda x: x == 'True',
         choices=[(True, 'Yes'), (False, 'No')],
     )
+
+
+class NoOperationForm(forms.Form):
+    pass